changes to itoa

+++ ft_isascii.c
@@ -12,7 +12,7 @@
 
 int	ft_isascii(int argument)
 {
-	if ((int) argument >= 0 && (int) argument <= 127)
+	if (argument >= 0 && argument <= 127)
 		return (1);
 	else
 		return (0);
